Move common heuristic code into a main class
