+++ sibyl/heuristics/arch.py
@@ -2,6 +2,8 @@
 
 from miasm2.analysis.binary import Container, ContainerUnknown
 
+from sibyl.heuristics.heuristic import Heuristic
+
 
 def container_guess(archinfo):
     """Use the architecture provided by the container, if any
@@ -16,41 +18,14 @@ def container_guess(archinfo):
     return {cont.arch: 1}
 
 
-class ArchHeuristic(object):
+class ArchHeuristic(Heuristic):
     """Provide heuristics to detect the architecture of a stream"""
 
     # Enabled passes
-    # passes are functions taking 'self' and returning a dict: candidates -> estimated probability
     heuristics = [
         container_guess,
     ]
 
     def __init__(self, stream):
+        super(ArchHeuristic, self).__init__()
         self.stream = stream
-        self._votes = None
-
-    def do_votes(self):
-        """Call heuristics and get back votes
-        Use a cumulative linear strategy for comparison
-        """
-        votes = {}
-        for heuristic in self.heuristics:
-            for name, vote in heuristic(self).iteritems():
-                votes[name] = votes.get(name, 0) + vote
-        self._votes = votes
-
-    @property
-    def votes(self):
-        """Cumulative votes for each candidates"""
-        if not self._votes:
-            self.do_votes()
-        return self._votes
-
-    def guess(self):
-        """Return the best candidate"""
-        sorted_votes = sorted(self.votes.iteritems(), key=lambda x:x[1])
-        if not sorted_votes:
-            # No solution
-            return False
-        best, _ = sorted_votes[-1]
-        return best
